+++ nuxt.config.ts
@@ -1,24 +1,23 @@
-import tailwindcss from "@tailwindcss/vite";
-
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
-  compatibilityDate: '2025-07-15',
+  compatibilityDate: "2025-07-15",
   devtools: { enabled: true },
-  css: ['./app/assets/css/main.css'],
-  vite: {
-    plugins: [
-      tailwindcss()
-    ]
-  },
+  types: ['@nuxtjs/i18n'],
   modules: [
-    '@nuxtjs/i18n'
+    '@nuxtjs/tailwindcss',
+    '@nuxt/image',
+    '@nuxt/fonts',
+    '@nuxt/icon',
+    '@pinia/nuxt',
+    '@nuxt/eslint',
+    '@nuxtjs/i18n',
   ],
   i18n: {
-    defaultLocale: 'id',
-    langDir: 'locales/',
+    defaultLocale: "id",
+    langDir: "locales/",
     locales: [
-      { code: 'en', file: 'en.json' },
-      { code: 'id', file: 'id.json' }
+      { code: "id", file: "id.json" },
+      { code: "en", file: "en.json" },
     ],
-  }
-})
+  },
+});