+++ tailwind.config.js
@@ -1,0 +1,18 @@
+/** @type {import('tailwindcss').Config} */
+export default {
+  // content: [],
+  theme: {
+    extend: {
+      fontFamily: {
+        inter: ["Inter", "sans-serif"],
+      },
+      colors: {
+        primary: "#0300B8",
+        secondary: "#4AF5C5",
+        tertiary: "#F4F4F4",
+        accent: "#CFCFCF",
+      },
+    },
+  },
+};
+