+++ CHANGELOG.md
@@ -6,3 +6,9 @@
 - First commit
 - Added chore features such as modules, packages, folders, and files
 - Added README.md and CHANGELOG.md 
+
+### V1.0.1
+- Created essentials components such as buttons, checkboxes, inputs, and header
+- Added localization folder
+- Added robux.png image in ./app/assets/images/robux.png
+- Added components.vue page for components sample